requirements.txt
certifi==2019.6.16
chardet==3.0.4
evernote3==1.25.13
httplib2==0.13.1
idna==2.8
oauth2==1.9.0.post1
oauthlib==3.1.0
pypandoc==1.4
python-dotenv==0.10.3
requests==2.22.0
requests-oauthlib==1.2.0
urllib3==1.25.3
